Separate edge client
